requirements/base.txt
Django==5.1
pillow==10.4.0
django-ckeditor-5==0.2.18
python-dotenv==1.1.1
